+++ src/app/crates/crates.component.spec.ts
@@ -1,6 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CratesComponent } from './crates.component';
+import { AppModule } from '../app.module';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { By } from '@angular/platform-browser';
 
 describe('CratesComponent', () => {
   let component: CratesComponent;
@@ -8,7 +11,9 @@ describe('CratesComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CratesComponent ]
+      declarations: [ CratesComponent ],
+      imports: [ AppModule ],
+      schemas: [ NO_ERRORS_SCHEMA ]
     })
     .compileComponents();
   }));
@@ -22,4 +27,15 @@ describe('CratesComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  it('click open button, calls open crate', () => {
+    const spy = spyOn(component, 'open');
+    const el = fixture.debugElement.query(By.css('#open')).nativeElement;
+
+    el.click();
+
+    expectAsync(spy).toBeResolved();
+  });
+
+  it('given reward')
 });