+++ src/environments/environment.prod.ts
@@ -4,7 +4,7 @@ export const environment = {
   url: 'https://3pg.xyz',
   discordURL: 'https://discord.com/invite/xRT6Fz5',
   stripePublicKey: 'pk_live_kyJ00K40kgFhHMYzO7Bc9YjV00sVDW6IlV',
-  version: 'v2.0.1a',
+  version: 'v2.0.3a',
   voteURLs: [
     '//discordbotlist.com/bots/525935335918665760/upvote'
   ]